Update README.md
I believe the AlarmMode in the items example should be a number instead of a string. In my install I found the mode would not render as expected when set to string (showed as a date), when swiched to number it displayed as expected. Also, in Channels section, mode is listed as a number.